requirements.txt
pytest~=8.4.1
aiohttp~=3.12.13
opencv-python~=4.9.0.80
numpy~=1.26.4
pillow~=11.2.1
pillow-avif-plugin
pydantic~=2.11.7
requests~=2.32.4
beautifulsoup4~=4.12.3
ezMM~=0.3.0
veritas~=0.1.12
Telethon~=1.40.0
PyYAML~=6.0.1
scraping~=0.0.3
config~=0.5.1
tweepy~=4.16.0
markdownify~=1.1.0
playwright~=1.53.0
tqdm~=4.66.3
util~=0.0.1
pytest-asyncio